discovered what `relative_paths` does...
If relative_paths is true then if the source is beneath the working directory
when the coverage is collected then the source file will have a relative path
recorded in the coverage data.  If relative_paths is false or the source is
not beneath the working directory then the path is still absolute.

in neither case is the path relative to the root of the project - it is still
relative to the environment-specific working directory.

Maybe if we have absolute paths everywhere then source file rewriting during
the `combine` step will work?